backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pymysql==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
pyjwt==2.9.0
bcrypt==3.2.2
passlib[bcrypt]